requirements.txt
filelock==3.12.0
Flask==2.3.2
Jinja2==3.1.2
PyYAML==6.0
selenium==4.11.2
tabulate==0.9.0
